refactor : History 클래스 domain 모듈로 이동
