+++ app/src/main/java/edu/nextstep/camp/calculator/HistoryAdapter.kt
@@ -1,10 +1,20 @@
 package edu.nextstep.camp.calculator
 
 import android.view.ViewGroup
+import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
-import edu.nextstep.camp.calculator.History.Companion.historyDiffUtil
+import edu.nextstep.camp.calculator.domain.History
 
-class HistoryAdapter : ListAdapter<History, HistoryViewHolder>(historyDiffUtil) {
+class HistoryAdapter : ListAdapter<History, HistoryViewHolder>(
+
+    object : DiffUtil.ItemCallback<History>() {
+        override fun areItemsTheSame(oldItem: History, newItem: History): Boolean =
+            oldItem.expression == newItem.expression
+
+        override fun areContentsTheSame(oldItem: History, newItem: History): Boolean =
+            oldItem == newItem
+    }) {
+
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int) =
         HistoryViewHolder.create(parent)
