+++ domain/src/main/java/edu/nextstep/camp/calculator/domain/History.kt
@@ -1,0 +1,6 @@
+package edu.nextstep.camp.calculator.domain
+
+data class History(
+    val expression: Expression,
+    val result: String
+)
